@@ -19,13 +19,14 @@
         "express": "^4.17.3",
         "http": "0.0.1-security",
         "jwt-decode": "^2.2.0",
-        "react": "^16.8.4",
-        "react-dom": "^16.8.4",
-        "react-redux": "^6.0.1",
-        "react-router-dom": "^5.0.0",
-        "react-scripts": "2.1.8",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2",
+        "react-redux": "^7.2.6",
+        "react-router-dom": "^6.2.1",
+        "react-scripts": "5.0.0",
         "redux": "^4.0.1",
         "redux-thunk": "^2.3.0",
+        "reselect": "^4.1.5",
         "typescript": "^4.5.5"
     },
     "browserslist": [
