@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component, Fragment, useState } from 'react';
 import { Link } from 'react-router-dom';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
@@ -15,123 +15,108 @@ import NotificationsIcon from '@material-ui/icons/Notifications';
 import FavoriteIcon from '@material-ui/icons/Favorite';
 import ChatIcon from '@material-ui/icons/Chat';
 // Redux
-import { connect } from 'react-redux';
+import { connect, useSelector } from 'react-redux';
 import { markNotificationsRead } from '../../redux/actions/userActions';
+import { notificationsSelector } from '../../redux/selectors/user';
 
-class Notifications extends Component {
-    state = {
-        anchorEl: null,
+function Notifications() {
+    const notifications = useSelector(notificationsSelector);
+    const [anchorEl, setAnchorEl] = useState(null);
+
+    const handleOpen = ({ target }) => {
+        setAnchorEl(target);
     };
-    handleOpen = (event) => {
-        this.setState({ anchorEl: event.target });
+
+    const handleClose = () => {
+        setAnchorEl(null);
     };
-    handleClose = () => {
-        this.setState({ anchorEl: null });
-    };
-    onMenuOpened = () => {
-        let unreadNotificationsIds = this.props.notifications
+
+    const onMenuOpened = () => {
+        let unreadNotificationsIds = notifications
             .filter((not) => !not.read)
             .map((not) => not.notificationId);
-        this.props.markNotificationsRead(unreadNotificationsIds);
+        markNotificationsRead(unreadNotificationsIds);
     };
-    render() {
-        const notifications = this.props.notifications;
-        const anchorEl = this.state.anchorEl;
 
-        dayjs.extend(relativeTime);
+    dayjs.extend(relativeTime);
+    let notificationsIcon;
 
-        let notificationsIcon;
-        if (notifications && notifications.length > 0) {
-            notifications.filter((not) => not.read === false).length > 0
-                ? (notificationsIcon = (
-                      <Badge
-                          badgeContent={
-                              notifications.filter((not) => not.read === false)
-                                  .length
-                          }
-                          color="secondary"
-                      >
-                          <NotificationsIcon />
-                      </Badge>
-                  ))
-                : (notificationsIcon = <NotificationsIcon />);
-        } else {
-            notificationsIcon = <NotificationsIcon />;
-        }
-        let notificationsMarkup =
-            notifications && notifications.length > 0 ? (
-                notifications.map((not) => {
-                    const verb = not.type === 'like' ? 'liked' : 'commented on';
-                    const time = dayjs(not.createdAt).fromNow();
-                    const iconColor = not.read ? 'primary' : 'secondary';
-                    const icon =
-                        not.type === 'like' ? (
-                            <FavoriteIcon
-                                color={iconColor}
-                                style={{ marginRight: 10 }}
-                            />
-                        ) : (
-                            <ChatIcon
-                                color={iconColor}
-                                style={{ marginRight: 10 }}
-                            />
-                        );
-
-                    return (
-                        <MenuItem
-                            key={not.createdAt}
-                            onClick={this.handleClose}
-                        >
-                            {icon}
-                            <Typography
-                                component={Link}
-                                color="default"
-                                variant="body1"
-                                to={`/users/${not.recipient}/scream/${not.screamId}`}
-                            >
-                                {not.sender} {verb} your scream {time}
-                            </Typography>
-                        </MenuItem>
-                    );
-                })
-            ) : (
-                <MenuItem onClick={this.handleClose}>
-                    You have no notifications yet
-                </MenuItem>
-            );
-        return (
-            <Fragment>
-                <Tooltip placement="top" title="Notifications">
-                    <IconButton
-                        aria-owns={anchorEl ? 'simple-menu' : undefined}
-                        aria-haspopup="true"
-                        onClick={this.handleOpen}
-                    >
-                        {notificationsIcon}
-                    </IconButton>
-                </Tooltip>
-                <Menu
-                    anchorEl={anchorEl}
-                    open={Boolean(anchorEl)}
-                    onClose={this.handleClose}
-                    onEntered={this.onMenuOpened}
-                >
-                    {notificationsMarkup}
-                </Menu>
-            </Fragment>
-        );
+    if (notifications && notifications.length > 0) {
+        notifications.filter((not) => not.read === false).length > 0
+            ? (notificationsIcon = (
+                  <Badge
+                      badgeContent={
+                          notifications.filter((not) => not.read === false)
+                              .length
+                      }
+                      color="secondary"
+                  >
+                      <NotificationsIcon />
+                  </Badge>
+              ))
+            : (notificationsIcon = <NotificationsIcon />);
+    } else {
+        notificationsIcon = <NotificationsIcon />;
     }
+
+    let notificationsMarkup =
+        notifications && notifications.length > 0 ? (
+            notifications.map((not) => {
+                const verb = not.type === 'like' ? 'liked' : 'commented on';
+                const time = dayjs(not.createdAt).fromNow();
+                const iconColor = not.read ? 'primary' : 'secondary';
+                const icon =
+                    not.type === 'like' ? (
+                        <FavoriteIcon
+                            color={iconColor}
+                            style={{ marginRight: 10 }}
+                        />
+                    ) : (
+                        <ChatIcon
+                            color={iconColor}
+                            style={{ marginRight: 10 }}
+                        />
+                    );
+
+                return (
+                    <MenuItem key={not.createdAt} onClick={handleClose}>
+                        {icon}
+                        <Typography
+                            component={Link}
+                            color="default"
+                            variant="body1"
+                            to={`/users/${not.recipient}/scream/${not.screamId}`}
+                        >
+                            {not.sender} {verb} ваші оголошення {time}
+                        </Typography>
+                    </MenuItem>
+                );
+            })
+        ) : (
+            <MenuItem onClick={handleClose}>У вас ще немає сповіщень</MenuItem>
+        );
+
+    return (
+        <Fragment>
+            <Tooltip placement="top" title="Notifications">
+                <IconButton
+                    aria-owns={anchorEl ? 'simple-menu' : undefined}
+                    aria-haspopup="true"
+                    onClick={handleOpen}
+                >
+                    {notificationsIcon}
+                </IconButton>
+            </Tooltip>
+            <Menu
+                anchorEl={anchorEl}
+                open={Boolean(anchorEl)}
+                onClose={handleClose}
+                onEntered={onMenuOpened}
+            >
+                {notificationsMarkup}
+            </Menu>
+        </Fragment>
+    );
 }
 
-Notifications.propTypes = {
-    markNotificationsRead: PropTypes.func.isRequired,
-    notifications: PropTypes.array.isRequired,
-};
-
-const mapStateToProps = (state) => ({
-    notifications: state.user.notifications,
-});
-
-export default connect(mapStateToProps, { markNotificationsRead })(
-    Notifications
-);
+export default Notifications;
