@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import withStyles from '@material-ui/core/styles/withStyles';
 // MUI Stuff
@@ -6,85 +6,83 @@ import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
 // Redux stuff
-import { connect } from 'react-redux';
+import { connect, useDispatch, useSelector } from 'react-redux';
 import { submitComment } from '../../redux/actions/dataActions';
+import { errorsSelector, loadingSelector } from '../../redux/selectors/ui';
+import { authenticatedSelector } from '../../redux/selectors/user';
 
 const styles = (theme) => ({
     ...theme,
 });
 
-class CommentForm extends Component {
-    state = {
+function CommentForm(props) {
+    const dispatch = useDispatch();
+    const errorsUI = useSelector(errorsSelector);
+    const loading = useSelector(loadingSelector);
+
+    const authenticated = useSelector(authenticatedSelector);
+
+    const [formData, setFormData] = useState({
         body: '',
-        errors: {},
-    };
+    });
+    const [errors, setErrors] = useState({});
 
-    componentWillReceiveProps(nextProps) {
-        if (nextProps.UI.errors) {
-            this.setState({ errors: nextProps.UI.errors });
+    useEffect(() => {
+        if (errorsUI !== errors) {
+            setErrors({ errors: errorsUI });
         }
-        if (!nextProps.UI.errors && !nextProps.UI.loading) {
-            this.setState({ body: '' });
+        if (!errorsUI && !loading) {
+            setFormData({ ...formData, body: '' });
         }
-    }
+    }, [errorsUI, loading]);
 
-    handleChange = (event) => {
-        this.setState({ [event.target.name]: event.target.value });
-    };
-    handleSubmit = (event) => {
-        event.preventDefault();
-        this.props.submitComment(this.props.screamId, {
-            body: this.state.body,
+    const handleChange = (event) => {
+        setFormData({
+            ...formData,
+            [event.target.name]: event.target.value,
         });
     };
 
-    render() {
-        const { classes, authenticated } = this.props;
-        const errors = this.state.errors;
+    const handleSubmit = (event) => {
+        event.preventDefault();
+        dispatch(
+            submitComment(props.screamId, {
+                body: formData.body,
+            })
+        );
+    };
 
-        const commentFormMarkup = authenticated ? (
-            <Grid item sm={12} style={{ textAlign: 'center' }}>
-                <form onSubmit={this.handleSubmit}>
-                    <TextField
-                        name="body"
-                        type="text"
-                        label="Comment on scream"
-                        error={errors.comment ? true : false}
-                        helperText={errors.comment}
-                        value={this.state.body}
-                        onChange={this.handleChange}
-                        fullWidth
-                        className={classes.textField}
-                    />
-                    <Button
-                        type="submit"
-                        variant="contained"
-                        color="primary"
-                        className={classes.button}
-                    >
-                        Submit
-                    </Button>
-                </form>
-                <hr className={classes.visibleSeparator} />
-            </Grid>
-        ) : null;
-        return commentFormMarkup;
-    }
+    const commentFormMarkup = authenticated ? (
+        <Grid item sm={12} style={{ textAlign: 'center' }}>
+            <form onSubmit={handleSubmit}>
+                <TextField
+                    name="body"
+                    type="text"
+                    label="Comment on scream"
+                    error={errors?.comment ? true : false}
+                    helperText={errors?.comment}
+                    value={formData.body}
+                    onChange={handleChange}
+                    fullWidth
+                    className={classes.textField}
+                />
+                <Button
+                    type="submit"
+                    variant="contained"
+                    color="primary"
+                    className={classes.button}
+                >
+                    Надіслати
+                </Button>
+            </form>
+            <hr className={classes.visibleSeparator} />
+        </Grid>
+    ) : null;
+    return commentFormMarkup;
 }
 
 CommentForm.propTypes = {
-    submitComment: PropTypes.func.isRequired,
-    UI: PropTypes.object.isRequired,
     classes: PropTypes.object.isRequired,
-    screamId: PropTypes.string.isRequired,
-    authenticated: PropTypes.bool.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-    UI: state.UI,
-    authenticated: state.user.authenticated,
-});
-
-export default connect(mapStateToProps, { submitComment })(
-    withStyles(styles)(CommentForm)
-);
+export default withStyles(styles)(CommentForm);
