@@ -1,88 +1,71 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React, { useEffect, useState } from 'react';
 import axios from 'axios';
-import Scream from '../components/scream/Scream';
+import Poster from '../components/scream/Poster';
 import StaticProfile from '../components/profile/StaticProfile';
 import Grid from '@material-ui/core/Grid';
 
-import ScreamSkeleton from '../util/ScreamSkeleton';
+import PosterSkeleton from '../util/PostersSkeleton';
 import ProfileSkeleton from '../util/ProfileSkeleton';
 
-import { connect } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { getUserData } from '../redux/actions/dataActions';
+import { loadingSelector, postersSelector } from '../redux/selectors/data';
 
-class user extends Component {
-    state = {
-        profile: null,
-        screamIdParam: null,
-    };
-    componentDidMount() {
-        const handle = this.props.match.params.handle;
-        const screamId = this.props.match.params.screamId;
+function User(props) {
+    const dispatch = useDispatch();
+    const posters = useSelector(postersSelector);
+    const loading = useSelector(loadingSelector);
+    const [profile, setProfile] = useState(null);
+    const [posterIdParam, setPosterIdParam] = useState(null);
 
-        if (screamId) this.setState({ screamIdParam: screamId });
+    useEffect(() => {
+        const handle = props?.handle;
+        const posterId = props?.posterId;
 
-        this.props.getUserData(handle);
+        if (posterId) setPosterIdParam(posterId);
+
+        dispatch(getUserData(handle));
         axios
             .get(`/user/${handle}`)
             .then((res) => {
-                this.setState({
-                    profile: res.data.user,
-                });
+                setProfile(res.data.user);
             })
             .catch((err) => console.log(err));
-    }
-    render() {
-        const { screams, loading } = this.props.data;
-        const { screamIdParam } = this.state;
+    }, []);
 
-        const screamsMarkup = loading ? (
-            <ScreamSkeleton />
-        ) : screams === null ? (
-            <p>No screams from this user</p>
-        ) : !screamIdParam ? (
-            screams.map((scream) => (
-                <Scream key={scream.screamId} scream={scream} />
-            ))
-        ) : (
-            screams.map((scream) => {
-                if (scream.screamId !== screamIdParam)
-                    return <Scream key={scream.screamId} scream={scream} />;
-                else
-                    return (
-                        <Scream
-                            key={scream.screamId}
-                            scream={scream}
-                            openDialog
-                        />
-                    );
-            })
-        );
+    const postersMarkup = loading ? (
+        <PosterSkeleton />
+    ) : posters === null ? (
+        <p>У цього користувача ще немає оголошень</p>
+    ) : !posterIdParam ? (
+        posters.map((scream) => (
+            <Poster key={scream.screamId} scream={scream} />
+        ))
+    ) : (
+        posters.map((scream) => {
+            if (scream.screamId !== posterIdParam)
+                return <Poster key={scream.screamId} scream={scream} />;
+            else
+                return (
+                    <Poster key={scream.screamId} scream={scream} openDialog />
+                );
+        })
+    );
 
-        return (
-            <Grid container spacing={16}>
-                <Grid item sm={8} xs={12}>
-                    {screamsMarkup}
-                </Grid>
-                <Grid item sm={4} xs={12}>
-                    {this.state.profile === null ? (
-                        <ProfileSkeleton />
-                    ) : (
-                        <StaticProfile profile={this.state.profile} />
-                    )}
-                </Grid>
+    return (
+        <Grid container spacing={16}>
+            <Grid item sm={8} xs={12}>
+                {postersMarkup}
             </Grid>
-        );
-    }
+            <Grid item sm={4} xs={12}>
+                {profile === null ? (
+                    <ProfileSkeleton />
+                ) : (
+                    <StaticProfile profile={profile} />
+                )}
+            </Grid>
+        </Grid>
+    );
 }
 
-user.propTypes = {
-    getUserData: PropTypes.func.isRequired,
-    data: PropTypes.object.isRequired,
-};
-
-const mapStateToProps = (state) => ({
-    data: state.data,
-});
-
-export default connect(mapStateToProps, { getUserData })(user);
+export default User;
