@@ -1,9 +1,9 @@
-import React, { Component, Fragment } from 'react';
+import React, { Fragment } from 'react';
 import { Link } from 'react-router-dom';
-import { connect } from 'react-redux';
-import PropTypes from 'prop-types';
+import { useSelector } from 'react-redux';
+import AppIcon from '../../images/icon.png';
 import MyButton from '../../util/MyButton';
-import PostScream from '../scream/PostScream';
+import PostPoster from '../scream/PostPoster';
 import Notifications from './Notifications';
 // MUI stuff
 import AppBar from '@material-ui/core/AppBar';
@@ -11,56 +11,49 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Button from '@material-ui/core/Button';
 // Icons
 import HomeIcon from '@material-ui/icons/Home';
+import { authenticatedSelector } from '../../redux/selectors/user';
+import { withStyles } from '@material-ui/core';
 
-class Navbar extends Component {
-    render() {
-        const { authenticated } = this.props;
-        return (
-            <AppBar>
-                <Toolbar className="nav-container">
-                    {authenticated ? (
-                        <Fragment>
-                            <PostScream />
-                            <Link to="/">
-                                <MyButton tip="Home">
-                                    <HomeIcon />
-                                </MyButton>
-                            </Link>
-                            <Notifications />
-                        </Fragment>
-                    ) : (
-                        <Fragment>
-                            <Button
-                                color="inherit"
-                                component={Link}
-                                to="/login"
-                            >
-                                Login
-                            </Button>
-                            <Button color="inherit" component={Link} to="/">
-                                Home
-                            </Button>
-                            <Button
-                                color="inherit"
-                                component={Link}
-                                to="/signup"
-                            >
-                                Signup
-                            </Button>
-                        </Fragment>
-                    )}
-                </Toolbar>
-            </AppBar>
-        );
-    }
-}
-
-Navbar.propTypes = {
-    authenticated: PropTypes.bool.isRequired,
-};
-
-const mapStateToProps = (state) => ({
-    authenticated: state.user.authenticated,
+const styles = (theme) => ({
+    ...theme,
 });
 
-export default connect(mapStateToProps)(Navbar);
+function Navbar({ classes }) {
+    const authenticated = useSelector(authenticatedSelector);
+
+    return (
+        <AppBar>
+            <Toolbar className="nav-container">
+                {authenticated ? (
+                    <Fragment>
+                        <PostPoster />
+                        <Link to="/">
+                            <MyButton tip="Home">
+                                <HomeIcon />
+                            </MyButton>
+                        </Link>
+                        <Notifications />
+                    </Fragment>
+                ) : (
+                    <div className="nav-wrapper">
+                        <img
+                            src={AppIcon}
+                            alt="ukraine"
+                            className={classes.logo}
+                            width="30px"
+                            height="30px"
+                        />
+                        <Button color="inherit" component={Link} to="/">
+                            Оголошення
+                        </Button>
+                    </div>
+                )}
+            </Toolbar>
+            <div className={classes.navbarTitle}>
+                <h1>Русский корабль иди нахуй!</h1>
+            </div>
+        </AppBar>
+    );
+}
+
+export default withStyles(styles)(Navbar);
