@@ -20,63 +20,60 @@ const styles = (theme) => ({
     },
 });
 
-class Comments extends Component {
-    render() {
-        const { comments, classes } = this.props;
-        return (
-            <Grid container>
-                {comments.map((comment, index) => {
-                    const { body, createdAt, userImage, userHandle } = comment;
-                    return (
-                        <Fragment key={createdAt}>
-                            <Grid item sm={12}>
-                                <Grid container>
-                                    <Grid item sm={2}>
-                                        <img
-                                            src={userImage}
-                                            alt="comment"
-                                            className={classes.commentImage}
+function Comments({ classes, comments }) {
+    return (
+        <Grid container>
+            {comments?.map((comment, index) => {
+                const { body, createdAt, userImage, userHandle } = comment;
+                return (
+                    <Fragment key={createdAt}>
+                        <Grid item sm={12}>
+                            <Grid container>
+                                <Grid item sm={2}>
+                                    <img
+                                        src={userImage}
+                                        alt="comment"
+                                        className={classes.commentImage}
+                                    />
+                                </Grid>
+                                <Grid item sm={9}>
+                                    <div className={classes.commentData}>
+                                        <Typography
+                                            variant="h5"
+                                            component={Link}
+                                            to={`/users/${userHandle}`}
+                                            color="primary"
+                                        >
+                                            {userHandle}
+                                        </Typography>
+                                        <Typography
+                                            variant="body2"
+                                            color="textSecondary"
+                                        >
+                                            {dayjs(createdAt).format(
+                                                'h:mm a, MMMM DD YYYY'
+                                            )}
+                                        </Typography>
+                                        <hr
+                                            className={
+                                                classes.invisibleSeparator
+                                            }
                                         />
-                                    </Grid>
-                                    <Grid item sm={9}>
-                                        <div className={classes.commentData}>
-                                            <Typography
-                                                variant="h5"
-                                                component={Link}
-                                                to={`/users/${userHandle}`}
-                                                color="primary"
-                                            >
-                                                {userHandle}
-                                            </Typography>
-                                            <Typography
-                                                variant="body2"
-                                                color="textSecondary"
-                                            >
-                                                {dayjs(createdAt).format(
-                                                    'h:mm a, MMMM DD YYYY'
-                                                )}
-                                            </Typography>
-                                            <hr
-                                                className={
-                                                    classes.invisibleSeparator
-                                                }
-                                            />
-                                            <Typography variabnt="body1">
-                                                {body}
-                                            </Typography>
-                                        </div>
-                                    </Grid>
+                                        <Typography variabnt="body1">
+                                            {body}
+                                        </Typography>
+                                    </div>
                                 </Grid>
                             </Grid>
-                            {index !== comments.length - 1 && (
-                                <hr className={classes.visibleSeparator} />
-                            )}
-                        </Fragment>
-                    );
-                })}
-            </Grid>
-        );
-    }
+                        </Grid>
+                        {index !== comments.length - 1 && (
+                            <hr className={classes.visibleSeparator} />
+                        )}
+                    </Fragment>
+                );
+            })}
+        </Grid>
+    );
 }
 
 Comments.propTypes = {
