@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component, Fragment, useState } from 'react';
 import withStyles from '@material-ui/core/styles/withStyles';
 import PropTypes from 'prop-types';
 import MyButton from '../../util/MyButton';
@@ -10,7 +10,7 @@ import DialogTitle from '@material-ui/core/DialogTitle';
 import DialogActions from '@material-ui/core/DialogActions';
 import DeleteOutline from '@material-ui/icons/DeleteOutline';
 
-import { connect } from 'react-redux';
+import { connect, useDispatch } from 'react-redux';
 import { deleteScream } from '../../redux/actions/dataActions';
 
 const styles = {
@@ -21,61 +21,49 @@ const styles = {
     },
 };
 
-class DeleteScream extends Component {
-    state = {
-        open: false,
-    };
-    handleOpen = () => {
-        this.setState({ open: true });
-    };
-    handleClose = () => {
-        this.setState({ open: false });
-    };
-    deleteScream = () => {
-        this.props.deleteScream(this.props.screamId);
-        this.setState({ open: false });
-    };
-    render() {
-        const { classes } = this.props;
+function DeleteScream({ classes, screamId }) {
+    const dispatch = useDispatch();
+    const [isOpen, setIsOpen] = useState(false);
 
-        return (
-            <Fragment>
-                <MyButton
-                    tip="Delete Scream"
-                    onClick={this.handleOpen}
-                    btnClassName={classes.deleteButton}
-                >
-                    <DeleteOutline color="secondary" />
-                </MyButton>
-                <Dialog
-                    open={this.state.open}
-                    onClose={this.handleClose}
-                    fullWidth
-                    maxWidth="sm"
-                >
-                    <DialogTitle>
-                        Are you sure you want to delete this scream ?
-                    </DialogTitle>
-                    <DialogActions>
-                        <Button onClick={this.handleClose} color="primary">
-                            Cancel
-                        </Button>
-                        <Button onClick={this.deleteScream} color="secondary">
-                            Delete
-                        </Button>
-                    </DialogActions>
-                </Dialog>
-            </Fragment>
-        );
-    }
+    const handleOpen = () => {
+        setIsOpen(true);
+    };
+    const handleClose = () => {
+        setIsOpen(false);
+    };
+    const deleteScream = () => {
+        dispatch(deleteScream(screamId));
+        setIsOpen(false);
+    };
+
+    return (
+        <Fragment>
+            <MyButton
+                tip="Видалити оголошення"
+                onClick={handleOpen}
+                btnClassName={classes.deleteButton}
+            >
+                <DeleteOutline color="secondary" />
+            </MyButton>
+            <Dialog open={isOpen} onClose={handleClose} fullWidth maxWidth="sm">
+                <DialogTitle>
+                    Ви впевненні, що хочете видалити це оголошення?
+                </DialogTitle>
+                <DialogActions>
+                    <Button onClick={handleClose} color="primary">
+                        Відмінити
+                    </Button>
+                    <Button onClick={deleteScream} color="secondary">
+                        Видалити
+                    </Button>
+                </DialogActions>
+            </Dialog>
+        </Fragment>
+    );
 }
 
 DeleteScream.propTypes = {
-    deleteScream: PropTypes.func.isRequired,
     classes: PropTypes.object.isRequired,
-    screamId: PropTypes.string.isRequired,
 };
 
-export default connect(null, { deleteScream })(
-    withStyles(styles)(DeleteScream)
-);
+export default withStyles(styles)(DeleteScream);
