@@ -1,6 +1,5 @@
-import React, { Component } from 'react';
+import React, { useEffect, useState } from 'react';
 import withStyles from '@material-ui/core/styles/withStyles';
-import PropTypes from 'prop-types';
 import AppIcon from '../images/icon.png';
 import { Link } from 'react-router-dom';
 
@@ -11,139 +10,118 @@ import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
 // Redux stuff
-import { connect } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { loginUser } from '../redux/actions/userActions';
+import { errorsSelector, loadingSelector } from '../redux/selectors/ui';
 
 const styles = (theme) => ({
     ...theme,
 });
 
-class login extends Component {
-    constructor() {
-        super();
-        this.state = {
-            email: '',
-            password: '',
-            errors: {},
-        };
-    }
-    componentWillReceiveProps(nextProps) {
-        if (nextProps.UI.errors) {
-            this.setState({ errors: nextProps.UI.errors });
+function Login({ classes, history }) {
+    const dispatch = useDispatch();
+    const loading = useSelector(loadingSelector);
+    const errorsUI = useSelector(errorsSelector);
+
+    const [errors, setErrors] = useState({});
+    const [formData, setFormData] = useState({
+        email: '',
+        password: '',
+    });
+
+    useEffect(() => {
+        if (errorsUI !== errors) {
+            setErrors({ errors: errorsUI });
         }
-    }
-    handleSubmit = (event) => {
+    }, [errorsUI]);
+
+    const handleSubmit = (event) => {
         event.preventDefault();
         const userData = {
-            email: this.state.email,
-            password: this.state.password,
+            email: formData.email,
+            password: formData.password,
         };
-        this.props.loginUser(userData, this.props.history);
+        dispatch(loginUser(userData, history));
     };
-    handleChange = (event) => {
-        this.setState({
+
+    const handleChange = (event) => {
+        setFormData({
+            ...formData,
             [event.target.name]: event.target.value,
         });
     };
-    render() {
-        const {
-            classes,
-            UI: { loading },
-        } = this.props;
-        const { errors } = this.state;
 
-        return (
-            <Grid container className={classes.form}>
-                <Grid item sm />
-                <Grid item sm>
-                    <img
-                        src={AppIcon}
-                        alt="ukraine"
-                        className={classes.image}
-                        width="100px"
-                        height="100px"
+    return (
+        <Grid container className={classes.form}>
+            <Grid item sm />
+            <Grid item sm>
+                <img
+                    src={AppIcon}
+                    alt="ukraine"
+                    className={classes.image}
+                    width="70px"
+                    height="70px"
+                />
+                <Typography variant="h2" className={classes.pageTitle}>
+                    Увійти
+                </Typography>
+                <form noValidate onSubmit={handleSubmit}>
+                    <TextField
+                        id="email"
+                        name="email"
+                        type="email"
+                        label="Email"
+                        className={classes.textField}
+                        helperText={errors?.email}
+                        error={errors?.email ? true : false}
+                        value={formData.email}
+                        onChange={handleChange}
+                        fullWidth
                     />
-                    <Typography variant="h2" className={classes.pageTitle}>
-                        Login
-                    </Typography>
-                    <form noValidate onSubmit={this.handleSubmit}>
-                        <TextField
-                            id="email"
-                            name="email"
-                            type="email"
-                            label="Email"
-                            className={classes.textField}
-                            helperText={errors.email}
-                            error={errors.email ? true : false}
-                            value={this.state.email}
-                            onChange={this.handleChange}
-                            fullWidth
-                        />
-                        <TextField
-                            id="password"
-                            name="password"
-                            type="password"
-                            label="Password"
-                            className={classes.textField}
-                            helperText={errors.password}
-                            error={errors.password ? true : false}
-                            value={this.state.password}
-                            onChange={this.handleChange}
-                            fullWidth
-                        />
-                        {errors.general && (
-                            <Typography
-                                variant="body2"
-                                className={classes.customError}
-                            >
-                                {errors.general}
-                            </Typography>
-                        )}
-                        <Button
-                            type="submit"
-                            variant="contained"
-                            color="primary"
-                            className={classes.button}
-                            disabled={loading}
+                    <TextField
+                        id="password"
+                        name="password"
+                        type="password"
+                        label="Password"
+                        className={classes.textField}
+                        helperText={errors?.password}
+                        error={errors?.password ? true : false}
+                        value={formData.password}
+                        onChange={handleChange}
+                        fullWidth
+                    />
+                    {errors.general && (
+                        <Typography
+                            variant="body2"
+                            className={classes.customError}
                         >
-                            Login
-                            {loading && (
-                                <CircularProgress
-                                    size={30}
-                                    className={classes.progress}
-                                />
-                            )}
-                        </Button>
-                        <br />
-                        <small>
-                            dont have an account ? sign up{' '}
-                            <Link to="/signup">here</Link>
-                        </small>
-                    </form>
-                </Grid>
-                <Grid item sm />
+                            {errors.general}
+                        </Typography>
+                    )}
+                    <Button
+                        type="submit"
+                        variant="contained"
+                        color="primary"
+                        className={classes.button}
+                        disabled={loading}
+                    >
+                        Увійти
+                        {loading && (
+                            <CircularProgress
+                                size={30}
+                                className={classes.progress}
+                            />
+                        )}
+                    </Button>
+                    <br />
+                    <small>
+                        Немає акаунта? Реєструйся <Link to="/signup">тут</Link>
+                    </small>
+                </form>
             </Grid>
-        );
-    }
+            <Grid item sm />
+        </Grid>
+    );
 }
 
-login.propTypes = {
-    classes: PropTypes.object.isRequired,
-    loginUser: PropTypes.func.isRequired,
-    user: PropTypes.object.isRequired,
-    UI: PropTypes.object.isRequired,
-};
-
-const mapStateToProps = (state) => ({
-    user: state.user,
-    UI: state.UI,
-});
-
-const mapActionsToProps = {
-    loginUser,
-};
-
-export default connect(
-    mapStateToProps,
-    mapActionsToProps
-)(withStyles(styles)(login));
+export default withStyles(styles)(Login);
