@@ -11,10 +11,10 @@ export const CLEAR_ERRORS = 'CLEAR_ERRORS';
 export const LOADING_DATA = 'LOADING_DATA';
 export const STOP_LOADING_UI = 'STOP_LOADING_UI';
 // Data reducer types
-export const SET_SCREAMS = 'SET_SCREAMS';
-export const SET_SCREAM = 'SET_SCREAM';
-export const LIKE_SCREAM = 'LIKE_SCREAM';
-export const UNLIKE_SCREAM = 'UNLIKE_SCREAM';
-export const DELETE_SCREAM = 'DELETE_SCREAM';
-export const POST_SCREAM = 'POST_SCREAM';
+export const SET_POSTERS = 'SET_POSTERS';
+export const SET_POSTER = 'SET_POSTER';
+export const LIKE_POSTER = 'LIKE_POSTER';
+export const UNLIKE_POSTER = 'UNLIKE_POSTER';
+export const DELETE_POSTER = 'DELETE_POSTER';
+export const POST_POSTER = 'POST_POSTER';
 export const SUBMIT_COMMENT = 'SUBMIT_COMMENT';
